fixed linting and testing
